+++ HealthPlanet/src/main/java/com/dongduk/HealthPlanet/controller/myPage/myPostListController.java
@@ -6,8 +6,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.dongduk.HealthPlanet.controller.Controller;
-import com.dongduk.HealthPlanet.model.Post;
-import com.dongduk.HealthPlanet.model.service.UserManager;
+import com.dongduk.HealthPlanet.domain.Post;
+import com.dongduk.HealthPlanet.service.UserManager;
 
 public class myPostListController implements Controller {
     @Override
