+++ HealthPlanet/src/main/java/com/dongduk/HealthPlanet/controller/myPage/postDeleteController.java
@@ -4,7 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.dongduk.HealthPlanet.controller.Controller;
-import com.dongduk.HealthPlanet.model.service.UserManager;
+import com.dongduk.HealthPlanet.service.UserManager;
 
 public class postDeleteController implements Controller {
     @Override
