+++ HealthPlanet/src/main/java/com/dongduk/HealthPlanet/controller/post/participatePostController.java
@@ -4,7 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.dongduk.HealthPlanet.controller.Controller;
-import com.dongduk.HealthPlanet.model.service.PostManager;
+import com.dongduk.HealthPlanet.service.PostManager;
 
 public class participatePostController implements Controller {
     @Override
