+++ HealthPlanet/src/main/java/com/dongduk/HealthPlanet/controller/post/postController.java
@@ -4,8 +4,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.dongduk.HealthPlanet.controller.Controller;
-import com.dongduk.HealthPlanet.model.Post;
-import com.dongduk.HealthPlanet.model.service.PostManager;
+import com.dongduk.HealthPlanet.domain.Post;
+import com.dongduk.HealthPlanet.service.PostManager;
 
 public class postController implements Controller {
     @Override
